requirements.txt
django==4.2.1
black==23.3
flake8==6.0.0
isort==5.12.0
pytest==7.3.2
djangorestframework==3.14.0
django_extensions==3.2.3
ipython==8.14.0
django-tree-queries==0.15.0
